requirements.txt
# Web Framework
fastapi>=0.135.0
uvicorn[standard]>=0.41.0
python-multipart>=0.0.9

# Data Validation & Settings
pydantic>=2.12.0
pydantic-settings>=2.0.0

# Database ORM
sqlmodel>=0.0.37
sqlalchemy>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Testing
pytest>=9.0.0
httpx>=0.27.0

# Linting & Formatting
ruff>=0.15.0
black>=26.0.0
